Fix reviewmanagement
Met Danny
