+++ src/ReviewManagementAPI/Dockerfile
@@ -1,0 +1,23 @@
+﻿FROM pitstop-dotnet-sdk-base:1.0 AS build-env
+WORKDIR /app
+
+# Copy necessary files and restore as distinct layer
+COPY *.csproj ./
+RUN dotnet restore
+
+# Copy everything else and build
+COPY . ./
+RUN dotnet publish -c Release -o out
+
+# Build runtime image
+FROM pitstop-dotnet-aspnet-base:1.0
+WORKDIR /app
+COPY --from=build-env /app/out .
+
+# Expose ports
+EXPOSE 5500/tcp
+ENV ASPNETCORE_URLS http://*:5500
+HEALTHCHECK --interval=30s --timeout=3s --retries=1 CMD curl --silent --fail http://localhost:5500/hc || exit 1
+
+# Start
+ENTRYPOINT ["dotnet", "Pitstop.ReviewManagementAPI.dll"]
