+++ src/ReviewManagementAPI/Models/ReviewModel.cs
@@ -1,7 +1,8 @@
-namespace ReviewManagmentAPI.Commands;
+namespace ReviewManagementAPI.Models;
 
-public class CreateReview
+public class Review
 {
+    public string Id { get; set; }
     public string Content { get; set; }
     public int Rating { get; set; }
     public string ReviewerName { get; set; }
