+++ src/ReviewManagementAPI/ReviewManagementAPI.http
@@ -1,0 +1,6 @@
+@ReviewManagementAPI_HostAddress = http://localhost:5194
+
+GET {{ReviewManagementAPI_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
