+++ src/ReviewManagementAPI/appsettings.Development.json
@@ -4,5 +4,8 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "Properties": {
+    "Application": "ReviewManagementAPI"
   }
 }
