+++ src/ReviewManagementAPI/appsettings.Production.json
@@ -1,0 +1,11 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "Properties": {
+    "Application": "ReviewManagementAPI",
+  }
+}
