--- src/ReviewManagmentAPI/Dockerfile
@@ -1,23 +1,0 @@
-﻿FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
-USER $APP_UID
-WORKDIR /app
-EXPOSE 8080
-EXPOSE 8081
-
-FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
-ARG BUILD_CONFIGURATION=Release
-WORKDIR /src
-COPY ["ReviewManagmentAPI/ReviewManagmentAPI.csproj", "ReviewManagmentAPI/"]
-RUN dotnet restore "ReviewManagmentAPI/ReviewManagmentAPI.csproj"
-COPY . .
-WORKDIR "/src/ReviewManagmentAPI"
-RUN dotnet build "ReviewManagmentAPI.csproj" -c $BUILD_CONFIGURATION -o /app/build
-
-FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "ReviewManagmentAPI.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "ReviewManagmentAPI.dll"]
