--- src/ReviewManagmentAPI/ReviewManagmentAPI.http
@@ -1,6 +1,0 @@
-@ReviewManagmentAPI_HostAddress = http://localhost:5251
-
-GET {{ReviewManagmentAPI_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
