--- src/ReviewManagmentAPI/Services/Interfaces/IReviewService.cs
@@ -1,11 +1,0 @@
-using ReviewManagmentAPI.Commands;
-using ReviewManagmentAPI.Models;
-
-namespace ReviewManagmentAPI.Services.Interfaces;
-
-public interface IReviewService
-{
-    Review AddReview(CreateReview command);
-    List<Review> GetAllReviews();
-    Review GetReviewById(string id);
-}
