+++ src/WebApp/appsettings.Development.json
@@ -16,6 +16,7 @@
   "APIServiceLocations": {
     "CustomerManagementAPI": "localhost:5100",
     "VehicleManagementAPI": "localhost:5000",
-    "WorkshopManagementAPI": "localhost:5200"
+    "WorkshopManagementAPI": "localhost:5200",
+    "ReviewManagementAPI": "localhost:5500"
   }
 }
