+++ src/WebApp/appsettings.Production.json
@@ -16,6 +16,7 @@
   "APIServiceLocations": {
     "CustomerManagementAPI": "customermanagementapi:5100",
     "VehicleManagementAPI": "vehiclemanagementapi:5000",
-    "WorkshopManagementAPI": "workshopmanagementapi:5200"
+    "WorkshopManagementAPI": "workshopmanagementapi:5200",
+    "ReviewManagementAPI": "reviewmanagementapi:5500"
   }
 }
